+++ apps/server/core/config.py
@@ -4,7 +4,6 @@ from pydantic_settings import BaseSettings, SettingsConfigDict
 class Settings(BaseSettings):
     PROJECT_NAME: str = "Inaam"
     DATABASE_URL: str = "postgresql+psycopg://postgres:password@localhost:5432/postgres"
-    SECRET_KEY: str
     JWT_SECRET_KEY: str
     CORS_ORIGINS: list[str] = ["http://localhost:3000"]